+++ src/core/Inflater.cpp
@@ -229,7 +229,7 @@ void Inflater::checkCapacity()
 
    // Increase
    double ratio = mCompIndex / (double)mCompLength;
-   int64 newLength = mUncompLength + max(4096l, (int64)(mUncompLength / ratio));
+   int64 newLength = mUncompLength + max(int64(4096), int64(mUncompLength / ratio));
    uint8* tmp = new uint8[newLength];
    if(tmp == nullptr)throw Exception("Cannot allocate memory!");
    memcpy(tmp, mUncompBytes, mUncompIndex);
